api for employee

See merge request green-academy/backend-green!37

+++ hr/src/main/java/org/green/hr/repository/TimeOffRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TimeOffRepository extends JpaRepository<TimeOff, Integer> {
+public interface TimeOffRepository extends JpaRepository<TimeOff, Long> {
 }
